ref: AddTrackPage: Separated UploadTrack Form into a different component for better readability
